refactor config code
