+++ setup/dlg_main_setup.cpp
@@ -1,11 +1,9 @@
-#include <gtk/gtk.h>
-#include <ibus.h>
 #include <string.h>
+#include <gtk/gtk.h>
 
 #include "dlg_macro_table.h"
 #include "dlg_main_setup.h"
-#include "engine_const.h"
-#include "utils.h"
+#include "config/unikey_config.h"
 
 static void init_dialog_controls(GtkBuilder*);
 static void input_method_combo_box_changed_cb(GtkComboBox*, gpointer);
@@ -13,19 +11,13 @@ static void output_charset_combo_box_changed_cb(GtkComboBox*, gpointer);
 static void update_config_toggle_cb(GtkToggleButton*, gpointer);
 static void macro_edit_button_cb(GtkButton*, gpointer);
 
-static GSettings* settings;
-
 GtkWidget* unikey_main_setup_dialog_new()
 {
     GtkBuilder* builder;
-    IBusBus* bus;
     GtkDialog* dlgMain;
 
-    ibus_init();
-    bus = ibus_bus_new();
-    g_signal_connect(bus, "disconnected", G_CALLBACK(gtk_main_quit), NULL);
-    settings = g_settings_new(UNIKEY_GSCHEMA_ID);
-    
+    ibus_unikey_config_init();
+
     builder = gtk_builder_new();
     gtk_builder_add_from_file(builder, UI_DATA_DIR "/setup-main.ui", NULL);
 
@@ -68,61 +60,59 @@ void init_dialog_controls(GtkBuilder* builder)
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "cbb_input_method"));
     g_signal_connect(wid, "changed", G_CALLBACK(input_method_combo_box_changed_cb), NULL);
     i = 0;
-    if (ibus_unikey_config_get_string(settings, CONFIG_INPUTMETHOD, &str))
-    {   
+    if (ibus_unikey_config_get_string(CONFIG_INPUTMETHOD, &str))
+    {
         for (; i < NUM_INPUTMETHOD; i++)
-        {   
-            if (strcasecmp(str, Unikey_IMNames[i]) == 0) break;
-        }   
+        {
+            if (strcasecmp(str, Unikey_IMNames[i]) == 0)
+                break;
+        }
+        g_free(str);
     }
     gtk_combo_box_set_active(GTK_COMBO_BOX(wid), i);
 
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "cbb_output_charset"));
     g_signal_connect(wid, "changed", G_CALLBACK(output_charset_combo_box_changed_cb), NULL);
     i = 0;
-    if (ibus_unikey_config_get_string(settings, CONFIG_OUTPUTCHARSET, &str))
-    {   
+    if (ibus_unikey_config_get_string(CONFIG_OUTPUTCHARSET, &str))
+    {
         for (; i < NUM_OUTPUTCHARSET; i++)
-        {   
-            if (strcasecmp(str, Unikey_OCNames[i]) == 0) break;
-        }   
+        {
+            if (strcasecmp(str, Unikey_OCNames[i]) == 0)
+                break;
+        }
+        g_free(str);
     }
     gtk_combo_box_set_active(GTK_COMBO_BOX(wid), i);
 
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "check_spellcheck"));
     g_signal_connect(wid, "toggled", G_CALLBACK(update_config_toggle_cb), (void*)CONFIG_SPELLCHECK);
-    if (!ibus_unikey_config_get_boolean(settings, CONFIG_SPELLCHECK, &b))
-        b = DEFAULT_CONF_SPELLCHECK;
+    ibus_unikey_config_get_boolean(CONFIG_SPELLCHECK, &b);
     gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(wid), b);
 
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "check_autorestorenonvn"));
     g_signal_connect(wid, "toggled", G_CALLBACK(update_config_toggle_cb), (void*)CONFIG_AUTORESTORENONVN);
-    if (!ibus_unikey_config_get_boolean(settings, CONFIG_AUTORESTORENONVN, &b))
-        b = DEFAULT_CONF_AUTONONVNRESTORE;
+    ibus_unikey_config_get_boolean(CONFIG_AUTORESTORENONVN, &b);
     gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(wid), b);
 
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "check_modernstyle"));
     g_signal_connect(wid, "toggled", G_CALLBACK(update_config_toggle_cb), (void*)CONFIG_MODERNSTYLE);
-    if (!ibus_unikey_config_get_boolean(settings, CONFIG_MODERNSTYLE, &b))
-        b = DEFAULT_CONF_MODERNSTYLE;
+    ibus_unikey_config_get_boolean(CONFIG_MODERNSTYLE, &b);
     gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(wid), b);
 
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "check_freemarking"));
     g_signal_connect(wid, "toggled", G_CALLBACK(update_config_toggle_cb), (void*)CONFIG_FREEMARKING);
-    if (!ibus_unikey_config_get_boolean(settings, CONFIG_FREEMARKING, &b))
-        b = DEFAULT_CONF_FREEMARKING;
+    ibus_unikey_config_get_boolean(CONFIG_FREEMARKING, &b);
     gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(wid), b);
 
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "check_macroenable"));
     g_signal_connect(wid, "toggled", G_CALLBACK(update_config_toggle_cb), (void*)CONFIG_MACROENABLED);
-    if (!ibus_unikey_config_get_boolean(settings, CONFIG_MACROENABLED, &b))
-        b = DEFAULT_CONF_MACROENABLED;
+    ibus_unikey_config_get_boolean(CONFIG_MACROENABLED, &b);
     gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(wid), b);
 
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "check_processwatbegin"));
     g_signal_connect(wid, "toggled", G_CALLBACK(update_config_toggle_cb), (void*)CONFIG_STANDALONEW);
-    if (!ibus_unikey_config_get_boolean(settings, CONFIG_STANDALONEW, &b))
-        b = DEFAULT_CONF_STANDALONEW;
+    ibus_unikey_config_get_boolean(CONFIG_STANDALONEW, &b);
     gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(wid), b);
 
     wid = GTK_WIDGET(gtk_builder_get_object(builder, "btn_macroedit"));
@@ -132,19 +122,19 @@ void init_dialog_controls(GtkBuilder* builder)
 void input_method_combo_box_changed_cb(GtkComboBox* cbb, gpointer user_data)
 {
     gint s = gtk_combo_box_get_active(cbb);
-    ibus_unikey_config_set_string(settings, CONFIG_INPUTMETHOD, Unikey_IMNames[s]);
+    ibus_unikey_config_set_string(CONFIG_INPUTMETHOD, Unikey_IMNames[s]);
 }
 
 void output_charset_combo_box_changed_cb(GtkComboBox* cbb, gpointer user_data)
 {
     gint s = gtk_combo_box_get_active(cbb);
-    ibus_unikey_config_set_string(settings, CONFIG_OUTPUTCHARSET, Unikey_OCNames[s]);
+    ibus_unikey_config_set_string(CONFIG_OUTPUTCHARSET, Unikey_OCNames[s]);
 }
 
 void update_config_toggle_cb(GtkToggleButton* btn, gpointer user_data)
 {
     gboolean b = gtk_toggle_button_get_active(btn);
-    ibus_unikey_config_set_boolean(settings, (gchar*)user_data, b);
+    ibus_unikey_config_set_boolean((gchar*)user_data, b);
 }
 
 void macro_edit_button_cb(GtkButton* btn, gpointer user_data)
