--- src/engine_const.h
@@ -1,26 +1,0 @@
-#ifndef __ENGINE_CONST_H__
-#define __ENGINE_CONST_H__
-
-#define UNIKEY_MACRO_FILE        ".ibus/unikey/macro"
-
-#define UNIKEY_GSCHEMA_ID        "org.freedesktop.ibus.engine.unikey"
-
-#define CONFIG_INPUTMETHOD       "input-method"
-#define CONFIG_OUTPUTCHARSET     "output-charset"
-#define CONFIG_SPELLCHECK        "spell-check"
-#define CONFIG_AUTORESTORENONVN  "auto-restore-non-vn"
-#define CONFIG_MODERNSTYLE       "modern-style"
-#define CONFIG_FREEMARKING       "free-marking"
-#define CONFIG_MACROENABLED      "macro-enabled"
-#define CONFIG_STANDALONEW       "standalone-w-as-uw"
-
-// DEFAULT options
-#define DEFAULT_CONF_SPELLCHECK         TRUE
-#define DEFAULT_CONF_AUTONONVNRESTORE   TRUE
-#define DEFAULT_CONF_MODERNSTYLE        FALSE
-#define DEFAULT_CONF_FREEMARKING        TRUE
-#define DEFAULT_CONF_MACROENABLED       FALSE
-#define DEFAULT_CONF_STANDALONEW        TRUE
-
-#endif
-
